Code refactored. Fixed the package conflict.
