+++ core/ui/src/main/java/co/yml/ytag/core/ui/templates/AppBar.kt
@@ -1,6 +1,6 @@
 @file:OptIn(ExperimentalMaterial3Api::class)
 
-package co.yml.ytag.ui.templates
+package co.yml.ytag.core.ui.templates
 
 import androidx.compose.foundation.layout.*
 import androidx.compose.material3.*
@@ -12,8 +12,8 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import co.yml.ytag.ui.R
-import co.yml.ytag.ui.theme.YTagTheme
+import co.yml.ytag.core.ui.R
+import co.yml.ytag.core.ui.theme.YTagTheme
 
 /**
  * Top app bars display information at the top of a screen.
