+++ core/ui/src/main/java/co/yml/ytag/core/ui/ytag/model/TagViewData.kt
@@ -1,4 +1,4 @@
-package co.yml.ytag.ui.ytag.model
+package co.yml.ytag.core.ui.ytag.model
 
 import androidx.compose.runtime.Composable
